proj-allotment-forbidden-batches: Change to camel case and refactor front end.
