+++ views/AUGSD/admin/portals/control/index.jade
@@ -36,7 +36,7 @@ block content
             li 
                 a(href='./control/switch-user') Switch User
             li 
-                a(href="#" onclick = "showInput()") Select forbidden batches for Project Allotment
+                a(href="#" onclick = "showInput()" id = "anchorForbiddenBatches") Select forbidden batches for Project Allotment
                 form(action="./control/project-allotment-forbidden-batches" method = "POST" style="display: none")#form
                     | Forbidden batches are 
                     if(years)
@@ -50,9 +50,17 @@ block content
                     button.mdl-button.mdl-js-button.mdl-button--raised.mdl-js-ripple-effect(type="submit") Submit
             script.
                 function showInput() {
-                    var x = document.getElementById("form");
-                    if(x.style.display === "none")
-                        x.style.display = "block";
+                    var select = document.getElementById("anchorForbiddenBatches");
+                    select.classList.toggle("selected")
+                    if(select.className == "selected")
+                        select.setAttribute("style", "color: navy");
                     else 
-                        x.style.display = "none";
+                        select.setAttribute("style", "color: none");
+                        
+                    var form = document.getElementById("form");
+
+                    if(form.style.display === "none")
+                        form.style.display = "block";
+                    else 
+                        form.style.display = "none";
                 }
